updated api

+++ app/src/main/java/com/example/shoppingcart/FoodAdapter.kt
@@ -74,7 +74,7 @@ class FoodAdapter(private var FoodItem: MutableList<Food>): RecyclerView.Adapter
                     if(check) {
                         val gQty: String? = sh.getString("pQty$getPos", "1")
                         val totalQty = gQty?.toInt()?.plus(etQty.text.toString().toInt())
-                        val totalPrice = (holder.itPrice.text as String).toInt().times(totalQty!!)
+                        val totalPrice = (holder.itPrice.text as String).toDouble().times(totalQty!!)
                         putString("pTitle$getPos", holder.itTitle.text.toString())
                         // price * quantity
                         putString("pPrice$getPos", totalPrice.toString())
@@ -83,7 +83,7 @@ class FoodAdapter(private var FoodItem: MutableList<Food>): RecyclerView.Adapter
                         check=false
                     }
                     else {
-                        val totalPrice = (holder.itPrice.text as String).toInt().times(etQty.text.toString().toInt())
+                        val totalPrice = (holder.itPrice.text as String).toDouble().times(etQty.text.toString().toInt())
                         putString("pTitle$gSize", holder.itTitle.text.toString())
                         putString("pPrice$gSize", totalPrice.toString())
                         putString("pQty$gSize", etQty.text.toString())
